requirement.txt
allure-pytest==2.9.45
allure-python-commons==2.9.45
atomicwrites==1.4.0
attrs==21.4.0
certifi==2022.6.15
charset-normalizer==2.0.12
colorama==0.4.5
demjson==2.2.4
genson==1.2.2
idna==3.3
iniconfig==1.1.1
jsonpath==0.82
jsonschema==4.17.3
packaging==21.3
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.9
pyrsistent==0.19.3
pytest==7.1.2
pytest-assume==2.4.3
PyYAML==6.0
requests==2.28.0
simplejson==3.17.6
six==1.16.0
tomli==2.0.1
urllib3==1.26.9
